requirements.txt
pandas~=1.5.3
numpy~=1.26.4
pyGCS~=0.4.2
opencv-python~=4.10.0.84
pillow~=10.3.0
dash~=2.17.1
matplotlib~=3.8.4
scipy~=1.11.4
scikit-learn~=1.4.2
plotly~=5.22.0
scikit-image~=0.24.0